+++ src/app/blog/[slug]/page.tsx
@@ -2,6 +2,9 @@ import { Fragment } from "react";
 import { getAllBlogs, getOneBlogArticle } from "@/services/blog.service";
 import { BlogArticleModel } from "@/models/models";
 import BlogArticleContainer from "@/components/blog-article-container/blog-article-container";
+import { calculateMetadata } from "@/utils/calculateMetadata";
+import { METADATA_INFOS } from "@/utils/constants";
+import { Metadata } from "next";
 // import { calculateMetadata } from "@/utils/calculateMetadata";
 
 // export async function generateMetadata({ params }: Props): Promise<Metadata> {
@@ -20,6 +23,11 @@ export function generateStaticParams() {
   return getAllBlogs().map((blog) => ({ slug: blog.slug }));
 }
 
+export async function generateMetadata({ params }: {params:{slug:string}}): Promise<Metadata> {
+  const slug = params.slug
+  const article = getOneBlogArticle(slug);
+  return calculateMetadata(article?.title ?? "Pas de titre", `/blog/${article?.slug}`, article?.cover as string, article?.content.slice(0,100)+'...')
+}
 export default async function Article({
   params,
 }: {