+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - [#21] 21-add-a-propos-page
+- [#20] 20-generate-metadata
 - [#10] 10-fix-navbar-blog-outline-in-blog-articles-part
 - [#6] 6-create-blog-section 
 - [#5] 5-improve-hero-aesthetic : Improve general aesthetic
@@ -19,4 +20,5 @@
 [#5]: https://github.com/willfynch/lalibrairiedesautoedites/issues/5
 [#6]: https://github.com/willfynch/lalibrairiedesautoedites/issues/6
 [#10]: https://github.com/willfynch/lalibrairiedesautoedites/issues/10
+[#20]: https://github.com/willfynch/lalibrairiedesautoedites/issues/20
 [#21]: https://github.com/willfynch/lalibrairiedesautoedites/issues/21